+++ src/components/helper-text/index.ts
@@ -1,0 +1,2 @@
+export type { HelperTextProps } from './types';
+export { HelperText } from './HelperText';