+++ src/components/textarea/types.ts
@@ -1,0 +1,5 @@
+import type { TextInputProps } from '../text-input';
+
+export interface TextareaProps extends Omit<TextInputProps, 'multiline'> {
+  rows?: number;
+}